Enlsip Documentation

Notations

The transpose of a given matrix M is written M'.

It corresponds to Adjoint operation in Julia, equivalent to transposition when using matrices with floats.

Structures

Structures to contain informations

Enlsip.Constraint — Type

Constraint

Struct used to represent the active constraints

Fields are the useful informations about active constraints at a point x :

cx : Vector of size t, contains values of constraints in current working set

A : Matrix of size t x t, jacobian matrix of constraints in current working set

scaling : Boolean indicating if internal scaling of cx and A is done

diag_scale : Vector of size t, contains the diagonal elements of the scaling matrix if internal scaling is done

The i-th element equals $\dfrac{1}{\|\nabla c_i(x)\|}$ for $i = 1,...,t$, which is the inverse of the length of A i-th row

Otherwise, it contains the length of each row in the matrix A

Enlsip.Iteration — Type

Iteration

Summarizes the useful informations about an iteration of the algorithm

x : Departure point of the iteration

p : Descent direction

rx : vector of size m, contains value of residuals at x

cx : vector of size l, constains value of constraints at x

t : Number of constraints in current working set (ie constraints considered active)

α : Value of steplength

λ : Vector of size t, containts Lagrange multipliers estimates

rankA : pseudo rank of matrix A, jacobian of active constraints

rankJ2 : pseudo rank of matrix J2, block extracted from J, jacobian of residuals

b_gn : right handside of the linear system solved to compute first part of p

d_gn : right handside of the linear system solved to compute second part of p

predicted_reduction : predicted linear progress

progress : reduction in the objective function

β : scalar used to estimate convergence factor

restart : indicate if current iteration is a restart step or no

first : indicate if current iteration is the first one or no

add : indicate if a constraint has been added to the working set

del : indicate if a constraint has been deleted from the working set

index_del : index of the constraint that has been deleted from working set (0 if no deletion)

code : Its value caracterizes the method used to compute the search direction p

1 represents Gauss-Newton method

-1 represents Subspace minimization

2 represents Newton method

nb_newton_steps : number of search direction computed using the method of Newton

Enlsip.WorkingSet — Type

WorkingSet

In ENLSIP, the working-set is a prediction of the set of active constraints at the solution

It is updated at every iteration thanks to a Lagrangian multipliers estimation

Fields of this structure summarize infos about the qualification of the constraints, i.e. :

q : number of equality constraints

t : number of constraints in current working set (all equalities and some inequalities considered to be active at the solution)

l : total number of constraints (i.e. equalities and inequalities)

active :

Vector of size l

first t elements are indeces of constraints in working set sorted in increasing order, other elements equal 0

inactive :

Vector of size l-q

first l-t elements are indeces of constraints not in working set sorted in increasing order, other elements equal 0

Evaluation functions structures

Enlsip.ConstraintsEval — Type

ConstraintsEval <: EvalFunc

Subtype of EvalFunc dedicated to the evalutation of constraints values and jacobian matrix.

Enlsip.EvalFunc — Type

EvalFunc(ctrl::Int64)

This structure is used to define functions evaluating residuals, constraints and corresponding jacobians.

Both functions for residuals and constraints must be written as follows and must not return any value (components are modified in the body of the function):

Example definition of an EvalFunc type function

function (h::EvalFunc)(x::Vector{Float64}, hx::Vector{Float64}, Jh::Matrix{Float64})
    if h.ctrl == 1 
        hx[:] = [...]
    elseif h.ctrl == 2
        Jh[:] = [...] # if supplied anatically
    end
    # The elseif block above could also be, if jacobian not supplied anatically
    # elseif h.ctrl == 2 h.ctrl = 0 end
    return
end

The ctrl field indicates what is computed (i.e. evalutation or jacobian) when calling a function of type EvalFunction.

ctrl = $1$ means the function h is evaluated at point x, (modifies in place vector hx)

ctrl = $2$ means the jacobian of h is computed at point x if jacobian is supplied anatically (then modifies in place matrix Jh)

if jacobian is not supplied anatically, ctrl is set to $0$ on return and jacobian is computed numerically.

Enlsip.ResidualsEval — Type

ResidualsEval <: EvalFunc

Subtype of EvalFunc dedicated to the evalutation of residuals values and jacobian matrix.

Functions

Enlsip.check_termination_criteria — Method

check_termination_criteria(iter::Iteration,prev_iter::Iteration,W::WorkingSet,active_C::Constraint,x,cx,rx_sum,∇fx,max_iter,nb_iter,ε_abs,ε_rel,ε_x,ε_c,error_code)

Equivalent Fortran77 routine : TERCRI

This functions checks if any of the termination criteria are satisfied

$\varepsilon_c,\varepsilon_x,\varepsilon_{rel}$ and $\varepsilon_{abs}$ are small positive values to test convergence.

There are convergence criteria (conditions 1 to 7) and abnormal termination criteria (conditions 8 to 12)

$\|c(x_k)\| < \varepsilon_c$ for constraints in the working set and all inactive constraints must be strictly positive

$\|A_k^T \lambda_k - \nabla f(x_k)\| < \sqrt{\varepsilon_{rel}}*(1 + \|\nabla f(x_k)\|)$

$\underset{i}{\min}\ \lambda_k^{(i)} \geq \varepsilon_{rel}*\underset{i}{\max}\ |\lambda_k^{(i)}|$

or $\underset{i}{\min}\ \lambda_k^{(i)} \geq \varepsilon_{rel}*(1+\|r(x_k)\|^2)$ if there is only one inequality

$\|d_1\|^2 \leq \varepsilon_x * \|x_k\|$

$\|r(x_k)\|^2 \leq \varepsilon_{abs}^2$

$\|x_k-x_{k-1}\| < \varepsilon_x * \|x_k\|$

$\dfrac{\sqrt{\varepsilon_{rel}}}{\|p_k\|} > 0.25$

number of iterations exceeds MAX_ITER

Convergence to a non feasible point

Second order derivatives not allowed by the user (TODO : not implemented yet)

Newton step fails or too many Newton steps done

The latest direction is not a descent direction to the merit function (TODO : not implemented yet)

Concitions 1 to 3 are necessary conditions.

This functions returns exit_code, an integer containing infos about the termination of the algorithm

0 if no termination criterion is satisfied

10000 if criterion 4 satisfied

2000 if criterion 5 satisfied

300 if criterion 6 satisfied

40 if criterion 7 satisfied

-2 if criterion 8 satisfied

-5 if criterion 11 satisfied

-9 if the search direction is computed with Newton method at least five times

-10 if not possible to satisfy the constraints

If multiple convergence criteria are satisfied, their corresponding values are added into exit_code.

exit_code != 0 means the termination of the algorithm

Enlsip.compute_steplength — Method

compute_steplength

Equivalent Fortran77 routine : STPLNG

Update the penalty weights and compute the steplength using the merit function psi

If search direction computed with method of Newton, an undamped step is taken, i.e. $\alpha =1$

On return

α : the computed steplength

w : vector of size l, containts the computed penalty constants

Enlsip.enlsip — Method

enlsip(x0,r,c,n,m,q,l;scaling = false,weight_code = 2,MAX_ITER = 100,ε_abs = 1e-10,ε_rel = 1e-3,ε_x = 1e-3,ε_c = 1e-3)

Main function for ENLSIP solver.

Must be called with the following arguments:

x0::Vector{Foat64} is the departure point

r is a function of type ResidualsEval to evaluate residuals values and jacobian

c is a function of type ConstraintsEval to evaluate constraints values and jacobian

n::Int64 is the number of parameters

m::Int64 is the number of residuals

q::Int64 is the number of equality constraints

l::Int64 is the total number of constraints (equalities and inequalities)

The following arguments are optionnal and have default values:

scaling::Bool

boolean indicating if internal scaling of constraints value and jacobian must be done or not

false by default

weight_code::Int64 is an int representing the method used to compute penality weights at each iteration

1 represents maximum norm method

2 (default value) represents euclidean norm method

MAX_ITER::Int64

int defining the maximum number of iterations

equals 100 by default

ε_abs, ε_rel, ε_x and ε_c

small positive scalars of type Float64 to test convergence

default are the recommended one by the authors, i.e.

ε_rel = ε_x = ε_c = 1e-3

ε_abs = 1e-10

Enlsip.first_lagrange_mult_estimate! — Method

first_lagrange_mult_estimate(A,λ,∇fx,cx,scaling_done,diag_scale,F)

Equivalent Fortran77 routine : MULEST

Compute first order estimate of Lagrange multipliers

Solves the system A' * λ_ls = ∇f(x) using QR factorisation of A' given by :

A'*P1 = Q1 * [R;0]

Then, computes estimates of lagrage multipliers by forming :

λ = λ_ls - inv(A*A') * cx

Parameters

A : nxt jacobian matrix of constraints in current working set

cx : vector of size t, contains evalutations of constraints in current working set

λ : vector of size t, represent the lagrange multipliers associated to current actives contraints

∇fx: vector of size n, equals the gradient vector of the objective function

scaling_done : Boolean indicating if internal scaling of contraints has been done or not

diag_scale : Vector of size t, contains the diagonal elements of the scaling matrix if internal scaling is done

The i-th element equals $\dfrac{1}{\|\nabla c_i(x)\|}$ for $i = 1,...,t$, which is the inverse of the length of A i-th row

Otherwise, it contains the length of each row in the matrix A

On return

Modifies in place the vector λ with the first order estimate of Lagrange multipliers.

Enlsip.gn_search_direction — Method

gn_search_direction(J,rx,cx,Q1,L11,P1,F_L11,rankA,t,τ,current_iter)

Equivalent Fortran77 routine : GNSRCH

Solves for y one of the compound systems :

    [L11;0] * y = b   
    J * y = -rx

or

    [R11;0] * y = Q2' * B
    J * y = -rx

Then, compute the search direction p = Q1 * y

If rankA = t, the first system is solved, otherwise, the second one is solved.

Parameters

J : mxn jacobian matrix of residuals

rx : residuals vector of size m

cx : active constraints vector of size t

Q1, L11, P1 : components of the LQ decomposition of active constraints jacobian matrix A*P1 = Q1 * [L11 ; 0]

Q1 orthogonal nxn orthogonal matrix

L11 txt lower triangular matrix

P1 txt permutation matrix

F_L11 : QRPivoted object containing infos about QR decomposition of matrix L11 such that

L11 * P2 = Q2 * [R11;0]

rankA : pseudo-rank of matrix A

τ : small positive value to compute the pseudo-rank of matrices

On return

p_gn : vector of size n, contains the computed search direction

F_J2 : QR decomposition of Matrix J2 defined in sub_search_direction

Enlsip.init_working_set — Method

init_working_set(cx,K,step,q,l)

Equivalent Fortran77 routine : INIALC

Compute the first working set by cheking which inequality constraints are strictly positive.

Then, initialize the penalty constants.

Parameters

cx : vector of size l, contains contraints evaluations

K : array of vectors, contains infos about penalty constants computed throughout the algorithm

step : object of type Iteration, containts infos about the current iteration, i.e. the first one when this function is called

q : number of equality constraints

l : total number of constraints

On return

first_working_set : WorkingSet object, contains infos about the first working set

Enlsip.new_point! — Method

new_point!(x,r,rx,c,cx,J,A,n,m,l)

Equivalent Fortran77 routine : NEWPNT

Compute in place the jacobians J and A corresponding to the residuals rx and the constraints cx evaluations respectively at current point x.

n is the number of parameters (size of x)

m is the number of residuals (size of rx)

lis the number of constraints (size of cx)

Enlsip.psi — Method

psi(x,α,p,r,c,w,m,l,t,active,inactive)

Compute and return the evaluation of the merit function at $(x+\alpha p,w)$ with current working set $\mathcal{W}$ and the set of inactive constraints $\mathcal{I}$

$\psi(x,w) = \dfrac{1}{2}\|r(x)\|^2 + \dfrac{1}{2}\sum_{i \in \mathcal{W}} w_ic_i(x)^2 + \dfrac{1}{2} \sum_{j \in \mathcal{I}} w_j\min(0,c_j(x))^2$

Enlsip.search_direction_analys — Function

search_direction_analys

Equivalent Fortran77 routine : ANALYS

Check if the latest step was sufficientlt good and eventually recompute the search direction by using either subspace minimization or the method of Newton

On return

error_code : integer indicating if there was an error if computations. In current version, errors can come from the method of Newton

Enlsip.sub_search_direction — Method

sub_search_direction(J1,rx,cx,Q1,L11,P1,F_L11,F_J2,n,t,rankA,dimA,dimJ2,code)

Equivalent Fortran77 routine : SUBDIR

Compute a search direction p by solving two triangular systems of equations.

First, for p1, either L11*p1 = -P1' * cx or R11*p1 = -Q2' * P1' * cx is solved.

Then for p2, R22*p2 = -Q3^T * [J1*p1 + rx] is solved.

[J1;J2] = J * Q1 * P2 where J is the jacobian matrix of residuals.

Finally, the search direction is computed by forming : p = Q1 * [p1 ; P3*p2]

Parameters

rx : residuals vector of size m

cx : active constraints vector of size t

Q1, L11, P1 : components of the LQ decomposition of active constraints jacobian matrix A*P1 = Q1 * [L11 ; 0]

Q1 orthogonal nxn orthogonal matrix

L11 txt lower triangular matrix

P1 txt permutation matrix

F_L11 : QRPivoted object containing infos about QR decomposition of matrix L11 such that

L11 * P2 = Q2 * [R11;0]

F_J2 : QRPivoted object containing infos about QR decomposition of matrix J2, last m-rankA columns of J*Q1

J2 * P3 = Q3 * [R22;0]

J1 first rankA columns of matrix J*Q1

n is the number of parameters

m is the number of residuals (size of rx)

t is the number of constraint in current working set

rankA : pseudo-rank of matrix A

dimA : number of columns of matrix R11 that should be used when R11*p1 = -Q2' * P1' * cx is solved

dimJ2 : number of columns of matrix R22 that should be used when R22*p2 = -Q3^T * [J1*p1 + rx] is solved

code : interger indicating which system to solve to compute p1

On return

p : vector of size n, contains the computed search direction

b : vector of size t, contains the right handside of the system solved to compute p1

d : vector of size m, contains the right handside of the system solved to compute p2

Enlsip.update_working_set — Method

Equivalent Fortran77 routine : WRKSET

First, an estimate the lagrange multipliers is computed.

If there are negative values among the multipliers computed, the constraint associated to the most negative multiplier is deleted from the working set.

Then, compute the search direction using Gauss-Newton method.

Parameters

W : represents the current working set (see WorkingSet for more details). Fields t, active and inactive may be modified when deleting a constraint

rx : vector of size m containing residuals evaluations

A : lxn jacobian matrix of constraints

J = mxn jacobian matrixe of residuals

C : represents constraints in current working set (see Constraint for more details)

∇fx : vector of size n, gradient vector of the objective function

p_gn : buffer vector of size n, represents the search direction

iter_k : Contains infos about the current iteration (see Iteration)

On return

P1, L11, Q1, F_L11 and F_J2 : QR decompositions used to solve linear systems when computing the search direction in sub_search_direction

The fields of iter_k related to the computation of the search direction are modified in place