Enlsip Documentation

Functions

Enlsip.enlsip — Function

enlsip(x0, r, c, n, m, q, l, scaling = false, weight_code = 2, MAX_ITER = 100)

Main function for ENLSIP solver.

Must be called with the following arguments:

x0::Vector{Foat64} is the departure point

r is a function of type ResidualsEval to evaluate residuals values and jacobian

c is a function of type ConstraintsEval to evaluate constraints values and jacobian

n::Int64 is the number of parameters

m::Int64 is the number of residuals

q::Int64 is the number of equality constraints

l::Int64 is the total number of constraints (equalities and inequalities)

The following arguments are optionnal and have default values:

scaling::Bool

boolean indicating if internal scaling of constraints value and jacobian must be done or not

false by default

weight_code::Int64 is an int representing the method used to compute penality weights at each iteration

1 represents maximum norm method

2 (default value) represents euclidean norm method

MAX_ITER::Int64

int defining the maximum number of iterations

equals 100 by default

name_output::String

string indicating the prefix of the .txt file produced at the end of the algorithm that contains informations about each iteration and termination of the algorithm

default name given is "output"

Evaluation functions structures

Enlsip.ConstraintsEval — Type

ConstraintsEval <: EvalFunc

Subtype of EvalFunc dedicated to the evalutation of constraints values and jacobian matrix.

Enlsip.EvalFunc — Type

EvalFunc(ctrl::Int64)

This structure is used to define functions evaluating residuals, constraints and corresponding jacobians.

Both functions for residuals and constraints must be written as follows and must not return any value (components are modified in the body of the function):

Example definition of an EvalFunc type function

function (h::EvalFunc)(x::Vector{Float64}, hx::Vector{Float64}, Jh::Matrix{Float64})
    if h.ctrl == 1 
        hx[:] = [...]
    elseif h.ctrl == 2
        Jh[:] = [...] # if supplied anatically
    end
    # The elseif block above could also be, if jacobian not supplied anatically
    # elseif h.ctrl == 2 h.ctrl = 0 end
    return
end

The ctrl field indicates what is computed (i.e. evalutation or jacobian) when calling a function of type EvalFunction.

ctrl = $1$ means the function h is evaluated at point x, (modifies in place vector hx)

ctrl = $2$ means the jacobian of h is computed at point x if jacobian is supplied anatically (then modifies in place matrix Jh)

if jacobian is not supplied anatically, ctrl is set to $0$ on return and jacobian is computed numerically.

Enlsip.ResidualsEval — Type

ResidualsEval <: EvalFunc

Subtype of EvalFunc dedicated to the evalutation of residuals values and jacobian matrix.